+++ src/table/consttable.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <stdbool.h>
-#include <stdint.h>
 
 #include "ccngen/enum.h"
 
@@ -14,7 +13,7 @@ typedef struct {
 typedef struct {
     enum BasicType ty;
     union {
-        int64_t intval;
+        int intval;
         double floatval;
     };
 } consttable_entry;
@@ -26,7 +25,7 @@ typedef struct {
 
 consttable *consttable_new(void);
 
-consttable_ref consttable_insert_int(consttable *self, int64_t intval);
+consttable_ref consttable_insert_int(consttable *self, int intval);
 
 consttable_ref consttable_insert_float(consttable *self, double floatval);
